Improve the transposition table interface a bit.
I'm still not sure why but the engine performs much weaker when using the
transposition table than when it is not. I have a feeling that I am not
using the transposition table correctly in the search. I have been up
until now just checking whether an entry for the hash of the current board
position exists in the table and if it does simply using the eval of that
entry as the eval of the current position. Quite possibly this in not
correct. Chess Wiki seems to indicate that we can only do this if the
eval is absolute, not an upper/lower bound due to alpha beta pruning.